changing events to events.service
service destinado a utilidades envolvendo eventos
