+++ src/app/page-restricted-area/page-manage-events/page-manage-events.page.ts
@@ -1,7 +1,7 @@
 import { GlobalConstantsService } from 'src/app/shared/services/global-constants.service';
 import { AlertController, ToastController } from '@ionic/angular';
 import { User } from 'src/app/shared/services/user';
-import { EventItem, EventSubscription } from 'src/app/shared/services/event';
+import { EventItem, EventSubscription } from 'src/app/shared/services/events.service';
 import { arrayRemove, arrayUnion, serverTimestamp } from '@angular/fire/firestore';
 import { Component, OnInit } from '@angular/core';
 import { AngularFirestore } from '@angular/fire/compat/firestore';
