+++ src/app/page-subscription/confirm-modal/confirm-modal.component.ts
@@ -5,7 +5,7 @@ import { Timestamp } from '@firebase/firestore-types';
 import { formatDate } from '@angular/common';
 import { fromUnixTime, isSameDay } from 'date-fns';
 import { MajorEventItem } from 'src/app/shared/services/major-event.service';
-import { EventItem } from 'src/app/shared/services/event';
+import { EventItem } from 'src/app/shared/services/events.service';
 import { ModalController } from '@ionic/angular';
 import { Observable } from 'rxjs';
 import { EnrollmentTypesService } from 'src/app/shared/services/enrollment-types.service';
