+++ src/app/page-subscription/info-modal/info-modal.component.ts
@@ -1,6 +1,6 @@
 import { ModalController } from '@ionic/angular';
 import { Component, Input, OnInit } from '@angular/core';
-import { EventItem } from 'src/app/shared/services/event';
+import { EventItem } from 'src/app/shared/services/events.service';
 import { parse } from 'twemoji-parser';
 import { DomSanitizer } from '@angular/platform-browser';
 import { DatesService } from 'src/app/shared/services/dates.service';
