+++ src/app/tab-calendar/components/calendar-list-view/calendar-list-view.component.ts
@@ -8,7 +8,7 @@ import { AngularFirestore } from '@angular/fire/compat/firestore';
 import { switchMap } from 'rxjs/operators';
 import { formatDate } from '@angular/common';
 import { DomSanitizer } from '@angular/platform-browser';
-import { EventItem } from 'src/app/shared/services/event';
+import { EventItem } from 'src/app/shared/services/events.service';
 import { trace } from '@angular/fire/compat/performance';
 import { Timestamp } from '@firebase/firestore-types';
 import { DatesService } from 'src/app/shared/services/dates.service';
